Introduce DeleteObjectId to directly delete from DB using ObjectID and not filename Used currently in MoveObject when we are overwriting another File / Dir

+++ melissi/actions/delete.py
@@ -2,6 +2,9 @@
 # DeleteObject
 # DeleteDir
 # DeleteFile
+# DeleteObjectId
+# DeleteObjectIdDroplet
+# DeleteObjectIdCell
 
 # standard modules
 import shutil
@@ -13,7 +16,30 @@ log = logging.getLogger("melissilogger")
 from melissi.actions import *
 
 class DeleteObject(WorkerAction):
-    """ lala """
+    """
+    To avoid racing conditions, where you delete a file and instantly
+    move / create another file into position, when the file / dir
+    still exists we issue a ModifyFile / ModifyDir and do nothing
+    more.
+
+    For example when firefox downloads a file follows the following
+    file creation pattern which is causing a racing condition.
+
+    For example let say that we download 'test.pdf'.
+
+    1. Firefox creates both test.pdf and test.pdf.part' in the
+    directory the file is to be saved
+
+    2. When download finishes, firefox deletes 'test.pdf' and moves
+    'test.pdf.part' to the new filename 'test.pdf'
+
+    3. Because of the deletion, notifier queues a DeleteFile action,
+    which probably doesn't get executed before the 'test.pdf.part' is
+    already named 'test.pdf'. As a result we delete the new 'test.pdf'
+    file and no 'test.pdf' or 'test.pdf.part' exists thereafter in the
+    folder
+    """
+
     def __init__(self, hub, filename, watchpath):
         super(DeleteObject, self).__init__(hub)
 
@@ -44,12 +70,23 @@ class DeleteObject(WorkerAction):
             log.debug("No record for [%s]" % self.unique_id)
             return
 
-        # delete from database
-        self._delete_from_db()
-        # delete from filesystem
-        self._delete_from_fs()
-        # notify server
-        return self._post_to_server()
+        if os.path.exists(pathjoin(self.watchpath, self.filename)):
+            if self._record.directory:
+                self._hub.queue.put(ModifyDir(self._hub,
+                                              self.filename,
+                                              self.watchpath)
+                                    )
+            else:
+                self._hub.queue.put(ModifyFile(self._hub,
+                                               self.filename,
+                                               self.watchpath)
+                                    )
+        else:
+            # delete from database
+            self._delete_from_db()
+
+            # notify server
+            return self._post_to_server()
 
     def _failure(self, result):
         log.error("Failure in delete %s" % result)
@@ -105,6 +142,70 @@ class DeleteFile(DeleteObject):
         except OSError, error_message:
             # ah, ignore
             pass
+
+    def _post_to_server(self):
+        uri = '%s/api/droplet/%s/' % (self._hub.config_manager.get_server(),
+                                  self._record.id)
+        d = self._hub.rest_client.delete(str(uri))
+        d.addErrback(self._failure)
+
+        return d
+
+
+class DeleteObjectId(WorkerAction):
+    def __init__(self, hub, objectid):
+        super(DeleteObjectId, self).__init__(hub)
+
+        self._objectid = objectid
+        self._record = False
+
+    @property
+    def unique_id(self):
+        return self._objectid
+
+    def exists(self):
+        return self._fetch_file_record(File__id=self._objectid)
+
+    def _execute(self):
+        self._record = self.exists()
+        if not self._record:
+            log.debug("No record with ib [%s]" % self.unique_id)
+            return
+
+        self._delete_from_db()
+        self._post_to_server()
+
+    def _failure(self, result):
+        log.error("Failure in delete objectid %s" % result)
+
+        raise RetryLater
+
+class DeleteObjectIdCell(DeleteObjectId):
+    def _delete_from_db(self):
+        # delete all children
+        for entry in self._dms.find(db.File,
+                                    db.File.filename.like(u'%s/%%' % self._record.filename),
+                                    db.WatchPath.id == self._record.watchpath_id,
+                                    db.WatchPath.id == db.File.watchpath_id
+                                    ):
+            self._dms.remove(entry)
+
+        # delete self
+        self._dms.remove(self._record)
+
+    def _post_to_server(self):
+        uri = '%s/api/cell/%s/' % (self._hub.config_manager.get_server(),
+                               self._record.id)
+        d = self._hub.rest_client.delete(str(uri))
+        d.addErrback(self._failure)
+
+        return d
+
+
+class DeleteObjectIdDroplet(DeleteObjectId):
+    def _delete_from_db(self):
+        # delete self
+        self._dms.remove(self._record)
 
     def _post_to_server(self):
         uri = '%s/api/droplet/%s/' % (self._hub.config_manager.get_server(),
